tenantcostclient: move code from kvcoord to tenantcostclient
Move the computeNetworkCost and computeWriteReplicationNetworkPaths functions
from the kvcoord package to the tenantcostclient package. This is part of a
larger change to refactor the OnResponseWait function on the tenant-side cost
controller. Keeping it in a separate commit makes the larger change easier to
review.

Epic: https://cockroachlabs.atlassian.net/browse/CC-28471

Release note: None
